Implement a central endpoint to select Overmind Instance
The new `--app` commandline flag is a
one-stop-shop to set all instance-specific values
from data in the frontend. This allows for a more
consistent and less error-prone way to select the
instance to work with. Notably this requires the
user to specify a frontend URL, which is easily
copy-pasted from the browser.