Ensure the risks get printed without truncation
bubbletea only outputs the visible portion of the View() based on the
terminal height. If the final View() with risks is taller than that, the
top parts get truncated before printing, screwing up scrollback. Formatting
and printing them in the cmdModel after the tea.Program has run avoids that
issue and provides a clean scrollback with all information.