Fix pre-command views
This creates a freeze/unfreezeViewMsg pair that can be used in conjunction
with the cliExecCommandModel to render the last View before a command
takes over the terminal into the scrollback buffer.

This gets rid of the janky skipView() implementation and should be much more
reliable as it reuses more of bubbletea's functionality by letting bubbletea
clear the View() by returning an empty string when frozen.