Fix initialisation order of tracing and logging
It turns out that PersistentPreRun are not always called in nested subcommands.
Since we already have PreRunSetup getting called everywhere, this commit
piggy-backs on that to setup logging and tracing _after_ the flags have
been parsed. This allows falling back to defaults when no environment
variable values have been provided.