Ensure that the view gets updated before fatalErrors are handled
As the channel processing is on a goroutine, while return values are
handled directly by tea, the view update from the last message on the
channel often loses the race against the CLI terminating from a fatalError.

This change moves all error reporting into the channel, so that messages
get processed in strict order, allowing tea to update the View a last
time before exiting.