Ignore errors to allow functional providers to be loaded into the CLI
This is a tough tradeoff between offering more functionality (providers
configured through data sources) and implementation complexity (lifting
a full terraform parser and evaluator to load data sources). For now we
err on the simple side and expose provider parse errors when `data`
references are used and continue processing as long as there is any
valid provider.