Default pterm output to stderr to not interfere with non-interactive output
Commands like submit-plan expect stdout to be clean to pass change-url
through when being used in the github action. This ensures that progress
information is a) displayed in CI and b) does not interfere with that
usage. Commands that are meant to be used interactively can provide
their own writer, like the `terraform` subcommands do.