Ensure that only valid args get passed to `terraform apply`
Native `terraform apply` does not accept `-var` or `-var-file` args when
applying a preplanned plan to avoid conflicts with the values stored in
said plan. As this is the main way we are running it internally, we also
make sure not to upset terraform in this case. To do so, this change
introduces a `applyArgsFromApplyArgs` method to strip that stuff out
before passing it into `RunApply`.