Query tracker rewrite (#919)
As part of testing some of the inconsistencies that we have been
noticing in the Gateway service, we determined that the query tracking
logic was faulty. At large scale it was possible for it to drop queries
and just generally behave weirdly. This isn't terribly surprising
looking at what the code used to look like once again this was one of
the first things that was written, and the code is extremely ugly and
uses a lot of go routines and synchronization logic that is really not
necessary.

This represents a complete rewrite from almost the ground up to
massively simplify the code base and make it much more reliable, and
less likely to be subject to race conditions.

https://github.com/user-attachments/assets/5ed46a3c-a045-4470-a558-51f4d02b6c82

GitOrigin-RevId: ab608791151c3c43934223a5403da882c72d2c30