Use item, not query, references when possible (#1002)
When processing revlinks, `GetReverseLinkedEdges` would create edges
from the `LinkedItemQueries` returned from revlink. Before this change
this would create a `IsQuery: true` reference. With this change, instead
a direct item reference is generated. See `GetReverseLinks` in revlink
to validate that only `Method: GET` queries are generated by revlink.

A proper integration test for this would be quite hairy. Instead to
expedite this fix this is adding a final sanity check that will notify
us through sentry if something else slips through.

This fixes #1000.

GitOrigin-RevId: 253662294650fc56b4ac91343f13061197904ff9