Fix loading snapshots with edges that have From queries (#1003)
This is in relation to #1000. Before the fix in #1002, edges that were
not processable were stored in the database. To avoid a data migration,
this commit fixes the data when it is loaded from the database.

To capture _which_ snapshot is affected, this change is a little bit
larger, passing through the snapId everywhere ParseSnapshotData is
called.

GitOrigin-RevId: dd543740e101e6e6e2a09956916b4f4a5e66a0ee