(fix) cli get change waits on 'in progress' and 'unspecified' (#1042)
re #1041

![image](https://github.com/user-attachments/assets/639b239b-9efe-4545-b070-649f07b800e2)
with the changes in where we do processing for a change. when we first
ask the status of the risk analysis, each step will be 0, ie unknown. We
should wait on this status of ChangeAnalysisStatus_STATUS_UNSPECIFIED
similar to ChangeAnalysisStatus_STATUS_INPROGRESS.

longer term GetChangeRisks will be removed, and we will ask the
change_timeline_v2 what is happening and this logic will be different.

GitOrigin-RevId: 295095de4be0675e8abfaab9006797f2e6163ca4