Implement a revlink.Checkpoint() endpoint for tests (#1046)
revlink ingestion is happening asynchronously. This is beneficial for
throughput in production, but also very annoying source of flakiness in
tests that rely on revlink being warmed up.

To allow tests to wait for revlink being fully committed, this adds a
`revlink.Checkpoint()` endpoint that waits until all currently in-flight
requests have made it into the database. The checkpoint endpoint is
forwarded into the Ingestor, which keeps track of everything that is
currently in flight.

### Limitations

* because of the use of `CondVar`, this will not work with multiple
revlink instances running
* When insertions are retried, a naive implementation with a single
target generation necessarily will ignore retries if a later insertion
has been committed.

Both of those are OK, since this is primarily for testing, where we
don't expect replication or retries to happen. The limitations are
documented on the endpoint to avoid surprises.

Fixes #1045

GitOrigin-RevId: 3527b4e6fcb75bbd044a74302bb05d67a0065387