Change revlink to directly return edges (#1078)
This removes a major user of `LinkedItemQueries` and moves us further
into a version of overmind without `sdp.TranslateLinksToEdges`.

Includes a touch of reformatting and CI cleanup

Note that this does rename the function, which is _technically_ a
breaking change, but since this is only used internally and all clients
are fixed at the same time, we can accept the risk of a few failing
requests during deployment.

Fixes #718

GitOrigin-RevId: 71a1bbfd103d04e0d7ffd7aa00b03c70c8044289