Remove a layer of channels from `QueryResultStream` (#1128)
QueryResultStream has a bunch of complicated channels/goroutines going
on that is not used by its callers.

This commit removes the waitgroup, channels and goroutines from
QueryResultStream. Instead the handlers get called immediately on
SendItem and SendError. This moves the need for synchronizing from the
stream to the handlers. To make that simple for tests, a new
RecordingQueryResultStream is implemented that is used by all tests for
inspecting the query results. The last remaining callers of
NewQueryResultStream use channels to report back and are safe.

This addresses part of #1126

GitOrigin-RevId: 3ddc5aa8f1fe4eea1ea61fc2f1e6f1fa22c4c83c