(fix) centralise token/auth telemetry (#1183)
To improve the [cli
stats](https://ui.honeycomb.io/overmind/environments/prod/board/zga4Skmt9AV/CLI-Stats)
this gathers setting all the relevant `omv.auth.*` attributes from a
token in a central method and calls it everywhere we receive a token.

The new method also sets the various bits of information from the token
in the context (`UserTokenContextKey`, `CustomClaimsContextKey`,
`CurrentSubjectContextKey`, and `AccountNameContextKey`) so that the
token only has to be parsed once and #1185 can pick up that information
from `ctx`.

GitOrigin-RevId: 15e4d5d65e63dcd3b97f8e27f79d37f46f688c70