use explicit blast propagation in dynamic adapters (#1852)
# Use Explicit Blast Propagations in Dynamic Adapters

Add the GCP resource type to SDP type conversions for the
[box.com](http://box.com) first milestone item types.

We will change now we identify the the SDP Item type during the linking
process. Instead of going from the attribute value and trying to figure
out what SDP type it is. We will start from the `FromSDPType`. That
means, we will update the `BlastPropagations` map to be more explicit by
defining the attribute key on the linked item. For example a compute
route item can have two network relations. One of them is the network
that the route belongs to (`network` is the attribute name). The other
is the network that it routes traffic to (`nextHopNetwork`). So, the
relation between the route object and these 2 different networks won't
be the same.
We need to differentiate this.
The new explicit blast propagation will look like:

```
var ExplicitBlastPropagations = map[shared.ItemType]map[string]Impact{
	ComputeRoute: {
		"network":        {}, // Network that the route belongs to
		"nextHopNetwork": {}, // Network that the route forwards traffic to, so the relationship will/may be different
	},
}
```

This will allow us to create the relations more accurately and make the
linking flow more straightforward with less guess work.

GitOrigin-RevId: dcb3c320fb4c93d9253210c45d0c38e7bcfe9a5d