Refactor feature flagging code (#1866)
As part of moving to GPT4.1 in a safe way, I want to use feature flags
to do it. However, I also want to report events of white space runaways
back to posthog. In order to do this, I need to be able to send
`Capture` events, so I needed to add this to the existing code.

After looking at the existing code and extrapolating forwards, how many
places in our code base are going to need to have feature flags in them
in the future, I decided that needing to pass the post hog client
everywhere was going to become painful. For example, if I was writing a
reasonably simple function that had a feature flag inside it, I would
need to pass the post hog client to that function. I would also need to
pass this in tests, et cetera.

This would mean that we would end up drilling this client into
essentially every single structure that we have or function signature.

In order to avoid this, I wanted to attach the client to the context.
This way if we're writing a test we can simply run the function and if
there is no post hog client, it will just return a default value. I have
also created test helpers so if we want to test the functionality of the
feature flag itself, we can do so.

I've also added a hell of a lot more documentation.

This could potentially be merged in its current form and then I could
build the new GPT 4.1 stuff on the back

GitOrigin-RevId: 538c3a19f674329751365ea1e02961eeafc18a63