feat: support terraform query map via id in gcp manual adapters (#1992)
# Review terraform mappings

We almost all the time use the `name` parameter from terraform resource.
In some cases the `name` parameter is the fully qualified name of the
resource. That means it will include project and/or location name as
well. However, our manual adapters are expecting the `name` query as a
simple string since the other params are handled via the scope of the
adapter.
So, we need to review existing manual adapters and make sure we take
handle the terraform `name` attribute correctly by extracting the simple
`name` (without the project or location information) if necessary.
Example terraform with fully qualified name:
[https://registry.terraform.io/providers/hashicorp/google/latest/docs/resources/google_service_account](https://registry.terraform.io/providers/hashicorp/google/latest/docs/resources/google_service_account)

GitOrigin-RevId: 7cde6e07e0827719764e5f51989470f5c1794d37