Eng 580 handle terraform mappings in dynamic adapters (#2017)
# Handle Terraform mappings in Dynamic Adapters

For some item types, the relevant Terraform resource does not provide
the expected argument for a GET request which is needed for terraform
mappings.
This is a case especially when an item is a child resource within a
project: i.e., a table in an instance or a key within a key ring.
In this cases, we should construct the relevant GET query params if the
query contains a slash, `/` which indicates that it can be a terraform
resource id.
Because in most cases (if not all), Terraform resources provides an `id`
attribute that holds all the information in a fully qualified name:
`projects/{{project}}/datasets/{{dataset}}/tables/{{table}}`
In the search method of an adapter we should check if the query is in
this format, and if it is, then extract the relevant information to
construct the required get query for the adapter.

We can use the helper function `ExtractPathParamsWithCount` introduced
with this PR.

[https://github.com/overmindtech/workspace/pull/1992](https://github.com/overmindtech/workspace/pull/1992)

GitOrigin-RevId: 0001af7107169c569372fd30bf5aec8b61077b42