test: add integration tests for (sub)network, and compute instance template (#2059)
# Add integration tests for gcp dynamic adapters

Let's add initial integration tests for:

* Compute Network
* Compute Subnetwork
* Compute Instance Template

We can use the default (sub)network resources, so we don't need to
create a new one for this tests.

Also, we can use the instance template that is created in instance
manager integration tests, so we don't need to create a new one here
either.

GitOrigin-RevId: 71ced4fa8c7c207823283499e15adf66340a3c20