feat, adding signal settings to sdp and api-server (#2195)
also does
https://linear.app/overmind/issue/ENG-934/implement-signals-settings

I was not around when the KV store was done for account config, it looks
weird how is uses atlas/sqlc

So my outstanding questions is how do you want to handle default values.
1. what are the default values i took them from the ticket
2. do we want to set the default values for all accounts in the DB, or
do what we currently do which is, unless the value is set in the DB, we
just give the system defaults.
3. when retrieving the signal settings if they dont exist, i save them
to the account. (mimicking the HCP behaviour)

GitOrigin-RevId: 7d528add288ffd489774a808dc5c4effe1e1ec74