Flush GKE prep work (#2278)
A number of refactorings and smaller fixes to prep for the last stretch
of getting gcp-sources over the line.

The most notable change in here is enabling Workload Identity Federation
on GKE, which will be required for the upcoming work of having srcman
manage all service accounts on the GCP/GKE side.

This was part of the ENG-1055 exploration.

GitOrigin-RevId: 4c77e97828d430f70863f16d0f400f9ed99f98ad