Fix helm chart naming and simply autoscaling (by removing it) (#2863)
As part of this PR I have removed the horizontal pod auto scaling
functionality from the helm chart since we have never used it and never
needed it and it adds complexity to customer deployment that harness
were asking about. I've also added a pod disruption budget so that the
pods won't get evicted too quickly and simplified to do just a simple
deployment without any autoscaling. Since we're limited by how quickly
we can query the API anyway there is essentially no chance that we're
going to max out the CPU on a source so as long as it's running we're
happy. Plus the user can still change the number of replicas in the helm
chart if they need to.

I also updated the name that the source users to use the user provided
name rather than the name of the chart. This is requires some
documentation updates as well.

Once this is complete, we're going to need to tag a new release i.e.

```
git tag k8s-source/v0.10.0
git push origin k8s-source/v0.10.0
```

GitOrigin-RevId: 5298f9c1431477c8db0cdcb7f85263047c52b403