Removed version from github artifact name (#3141)
By doing this it means that we can use the `/latest` URL to download the
latest version of the CLI. This means that the plugin for Env0 can
become much simpler as it will no longer need to query the latest
version from the Github API. This also means that in ENG-1872 we could
use the `binaries` option to download the CLI. since we could hardcode a
URL that we won't need to perpetually update.

I'm pretty sure this won't affect anything since all of the packages
that we publish are covered by other goreleaser configs.

Once this is merged we will need to:
- Release the CLI
- Update the plugin for Env0 to use the `/latest` URL

<!-- CURSOR_SUMMARY -->
---

> [!NOTE]
> Removes the version segment from the GoReleaser archive name template
for the CLI artifacts.
>
> <sup>Written by [Cursor
Bugbot](https://cursor.com/dashboard?tab=bugbot) for commit
8b3bbab9667ab474601886a8bbe02d03e4d45896. Configure
[here](https://cursor.com/dashboard?tab=bugbot).</sup>
<!-- /CURSOR_SUMMARY -->

GitOrigin-RevId: f6af9e531546b0e94c862f4de9915b59dd4d267f